requirements.txt
requests
google-cloud-storage
Flask
gunicorn